Revert "Auto download"
